Right before character test create read
The modifications in this commit questioned if some classes really needed to have a reference class that already refers to it. I.E.: DaggerheartCharacter.java and Inventory.java, for the use of the app an inventory cannot exist without a character, thus it has no need to refer to a character.
Otherwise, the addition of DTO(data transfer object) ensures security and ease of creating a object without the need of its id.